Extract runBundledAgent helper to deduplicate agent/review CLI commands
